+++ fidget-wgpu/src/pixel/mod.rs
@@ -5,7 +5,7 @@
 use crate::{
     Gpu, RegPipeline, RenderShape, TAPE_DATA_CAPACITY, TapeWord,
     buf::{
-        BufferItemCount, BufferSizeError, BufferType, FlexBuffer, buffer_ro,
+        BufferSizeError, BufferType, FlexBuffer, ReadBuffer, buffer_ro,
         buffer_rw,
     },
     shaders, tag,
@@ -413,9 +413,6 @@ tag!(pub PixelBufferTag, RawDistancePixel, ImageSize, STORAGE | COPY_SRC | COPY_
 ///
 /// This object is constructed by [`Context::buffers`] and may only be used with
 /// that particular [`Context`].
-///
-/// A successfully constructed (or resized) `Buffers` object also guarantees
-/// infallible construction of an [`ImageReadBuffer`] object of the same size.
 pub struct Buffers {
     /// Image render size
     ///
@@ -438,15 +435,6 @@ pub struct Buffers {
     /// Pixel data
     pixels: FlexBuffer<PixelBufferTag>,
 
-    /// Query set for timestamps
-    ///
-    /// This must be present if and only if the parent context has timestamps
-    /// enabled (per [`Context::has_timestamps`])
-    timestamps: Option<wgpu::QuerySet>,
-
-    /// Buffer into which we resolve the timestamp query
-    ts_buf: wgpu::Buffer,
-
     /// Cached bind groups
     bind_groups: BindGroups,
 }
@@ -455,7 +443,7 @@ impl Buffers {
     /// Builds a new set of buffers with a default size
     ///
     /// It is expected that these will be resized before being used
-    fn new(device: &wgpu::Device, has_timestamps: bool) -> Self {
+    fn new(device: &wgpu::Device) -> Self {
         // The config buffer is statically sized, so we can check it here
         static_assertions::const_assert!(
             (std::mem::size_of::<Config>()
@@ -487,26 +475,8 @@ impl Buffers {
         let pixels =
             FlexBuffer::new(device, "pixels".to_string(), image_size).unwrap();
 
-        let ts_buf = device.create_buffer(&wgpu::BufferDescriptor {
-            label: Some("ts"),
-            size: 2 * std::mem::size_of::<u64>() as u64,
-            usage: wgpu::BufferUsages::QUERY_RESOLVE
-                | wgpu::BufferUsages::COPY_SRC,
-            mapped_at_creation: false,
-        });
-
         let tile64 = TileBuffers::new(device, render_size).unwrap();
         let tile8 = TileBuffers::new(device, render_size).unwrap();
-
-        let timestamps = if has_timestamps {
-            Some(device.create_query_set(&wgpu::QuerySetDescriptor {
-                label: Some("timestamp query set"),
-                ty: wgpu::QueryType::Timestamp,
-                count: 2,
-            }))
-        } else {
-            None
-        };
 
         Self {
             config_buf,
@@ -515,22 +485,8 @@ impl Buffers {
             tile64,
             tile8,
             pixels,
-            timestamps,
-            ts_buf,
             bind_groups: Default::default(),
         }
-    }
-
-    /// Returns image buffer size (in bytes)
-    fn image_buf_size(image_size: ImageSize) -> usize {
-        image_size
-            .item_count()
-            // Convert from GeometryPixel item count to bytes
-            .checked_mul(std::mem::size_of::<RawDistancePixel>())
-            .unwrap()
-            // Allocate an extra 16 bytes for timestamp queries
-            .checked_add(16)
-            .unwrap()
     }
 
     /// Returns the number of bytes in the `tile_tapes` buffer
@@ -557,10 +513,6 @@ impl Buffers {
     /// Resizes to render the target image size
     ///
     /// Internal buffers are resized to fit (only getting larger)
-    ///
-    /// This function also checks that the size is appropriate for an
-    /// [`ImageReadBuffer`] (though we do not store such an object), so that
-    /// later functions can resize it infallibly.
     fn set_image_size(
         &mut self,
         device: &wgpu::Device,
@@ -574,8 +526,6 @@ impl Buffers {
             tile8,
             config_buf: _,
             pixels,
-            timestamps: _,
-            ts_buf: _,
             bind_groups,
         } = self;
         // Clear our cached bind groups if the image sizes is changing
@@ -607,15 +557,6 @@ impl Buffers {
                 buf: BufferName::Pixel,
                 err,
             })?;
-
-        // Check that we can build an `ImageReadBuffer` of the appropriate
-        // size (even though they are stored separately)
-        ImageReadArrayBuffer::check_size(Self::image_buf_size(image_size))
-            .map_err(|err| BuffersError {
-                buf: BufferName::Image,
-                err,
-            })?;
-
         Ok(())
     }
 
@@ -629,8 +570,6 @@ impl Buffers {
             tile64,
             tile8,
             pixels,
-            timestamps: _,
-            ts_buf,
             bind_groups: _,
         } = self;
         config_buf.size()
@@ -638,7 +577,6 @@ impl Buffers {
             + tile64.capacity()
             + tile8.capacity()
             + pixels.capacity()
-            + ts_buf.size()
     }
 
     /// Returns total active size (in bytes)
@@ -651,8 +589,6 @@ impl Buffers {
             tile64,
             tile8,
             pixels,
-            timestamps: _,
-            ts_buf,
             bind_groups: _,
         } = self;
         config_buf.size()
@@ -660,7 +596,6 @@ impl Buffers {
             + tile64.size()
             + tile8.size()
             + pixels.size_bytes()
-            + ts_buf.size()
     }
 
     /// Returns a handle to the image storage buffer
@@ -669,7 +604,7 @@ impl Buffers {
     /// [`RawDistancePixel`] image data without copying to the CPU.  It requires
     /// a exclusive borrow of the `Buffers` object (and then extends that
     /// lifetime) so that other callers can't simultaneously touch the buffer.
-    pub fn image_storage_buffer(&mut self) -> &FlexBuffer<PixelBufferTag> {
+    pub fn output(&mut self) -> &FlexBuffer<PixelBufferTag> {
         &self.pixels
     }
 }
@@ -858,104 +793,6 @@ impl BindGroups {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-tag!(ImageReadTag, u8, usize, COPY_DST | MAP_READ);
-type ImageReadArrayBuffer = FlexBuffer<ImageReadTag>;
-
-/// Buffer for reading data back from the GPU
-///
-/// This object is constructed by [`Context::image_buffer`] and may only be used
-/// with that particular [`Context`].
-///
-/// Once mapped, this is wrapped by a [`MappedImage`]
-pub struct ImageReadBuffer {
-    /// Image render size
-    image_size: ImageSize,
-
-    /// Result buffer that can be read back from the CPU
-    ///
-    /// This is mostly image pixels (as [`RawDistancePixel`] values), but also
-    /// contains two trailing `u64` values for timestamps.
-    buffer: ImageReadArrayBuffer,
-}
-
-impl ImageReadBuffer {
-    fn new(device: &wgpu::Device, name: String) -> Self {
-        let image_size = 64.into();
-        Self {
-            image_size,
-            buffer: ImageReadArrayBuffer::new(
-                device,
-                name,
-                Buffers::image_buf_size(image_size),
-            )
-            .expect("64 should always be a valid size"),
-        }
-    }
-
-    fn grow_to_fit(
-        &mut self,
-        device: &wgpu::Device,
-        image_size: ImageSize,
-    ) -> Result<(), BufferSizeError> {
-        self.image_size = image_size;
-        self.buffer
-            .grow_to_fit(device, Buffers::image_buf_size(image_size))
-    }
-}
-
-/// Handle to a mapped image, which unmaps the image when dropped
-pub struct MappedImage<'a> {
-    image: &'a ImageReadBuffer,
-    slice: wgpu::BufferSlice<'a>,
-
-    /// Nanoseconds per tick, for resolving timestamps
-    ns_per_tick: Option<f32>,
-}
-
-impl Drop for MappedImage<'_> {
-    fn drop(&mut self) {
-        self.image.buffer.data().unmap();
-    }
-}
-
-impl MappedImage<'_> {
-    /// Returns the image's data
-    pub fn image(&self) -> Image {
-        // Get the pixel-populated image
-        let result = <[RawDistancePixel]>::ref_from_bytes(
-            &self.slice.get_mapped_range()[..self.image_bytes()],
-        )
-        .unwrap()
-        .to_owned();
-        Image::build(result, self.image.image_size).unwrap()
-    }
-
-    /// Returns the time spent in the compute pass
-    ///
-    /// This may be 0 on platforms which advertise `TIMESTAMP_QUERY` but do not
-    /// actually populate timestamps, and will be `None` if the context does not
-    /// have `TIMESTAMP_QUERY` enabled.
-    pub fn time(&self) -> Option<std::time::Duration> {
-        self.ns_per_tick.map(|ns_per_tick| {
-            let slice = self.slice.get_mapped_range();
-            let ts =
-                <[u64]>::ref_from_bytes(&slice[self.image_bytes()..]).unwrap();
-            std::time::Duration::from_nanos(
-                (ts[1].saturating_sub(ts[0]) as f64 * ns_per_tick as f64)
-                    as u64,
-            )
-        })
-    }
-
-    fn image_bytes(&self) -> usize {
-        (self.image.image_size.width() as usize)
-            * (self.image.image_size.height() as usize)
-            * std::mem::size_of::<RawDistancePixel>()
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
 tag!(TilesBufferTag, u32, usize, STORAGE | COPY_DST | INDIRECT);
 tag!(ValuesBufferTag, u32, ImageSize, STORAGE | COPY_DST);
 
@@ -1087,7 +924,6 @@ impl<const N: usize> TileBuffers<N> {
 /// Context for 2D (distance field) rendering
 pub struct Context {
     gpu: Gpu,
-    has_timestamps: bool,
 
     /// Bind group layout for the common bind group (used by all stages)
     common_bind_group_layout: wgpu::BindGroupLayout,
@@ -1117,17 +953,6 @@ impl Context {
     /// If render timestamps are desirable, then the device should be
     /// initialized with [`wgpu::Features::TIMESTAMP_QUERY`].
     pub fn new(gpu: &Gpu) -> Self {
-        let has_timestamps = gpu
-            .device
-            .features()
-            .contains(wgpu::Features::TIMESTAMP_QUERY);
-        if !has_timestamps {
-            log::warn!(
-                "WGPU device is missing `TIMESTAMP_QUERY`; \
-                 timestamps are disabled"
-            );
-        }
-
         // Create bind group layout and bind group
         let common_bind_group_layout = gpu.device.create_bind_group_layout(
             &wgpu::BindGroupLayoutDescriptor {
@@ -1170,7 +995,6 @@ impl Context {
 
         Self {
             gpu: gpu.clone(),
-            has_timestamps,
             common_bind_group_layout,
             vars_bind_group_layout,
             root_ctx,
@@ -1187,12 +1011,7 @@ impl Context {
     /// when passed into any of the runner functions (e.g. [`run`](Self::run) or
     /// [`submit`](Self::submit)).
     pub fn buffers(&self) -> Buffers {
-        Buffers::new(&self.gpu.device, self.has_timestamps)
-    }
-
-    /// Returns an [`ImageReadBuffer`] to read from a [`Buffers`] object
-    pub fn image_buffer(&self) -> ImageReadBuffer {
-        ImageReadBuffer::new(&self.gpu.device, "image".to_owned())
+        Buffers::new(&self.gpu.device)
     }
 
     /// Renders the image, with a blocking wait to read pixel data from the GPU
@@ -1203,7 +1022,7 @@ impl Context {
         &self,
         shape: &RenderShape,
         buffers: &mut Buffers,
-        out: &mut ImageReadBuffer,
+        out: &mut ReadBuffer<PixelBufferTag>,
         settings: RenderConfig,
     ) -> Result<Image, SubmitError> {
         self.run_with_vars(shape, &Default::default(), buffers, out, settings)
@@ -1218,23 +1037,21 @@ impl Context {
         shape: &RenderShape,
         vars: &ShapeVars<f32>,
         buffers: &mut Buffers,
-        out: &mut ImageReadBuffer,
+        out: &mut ReadBuffer<PixelBufferTag>,
         settings: RenderConfig,
     ) -> Result<Image, SubmitError> {
         self.submit_with_vars(shape, vars, buffers, &settings)?;
-        let image = self.map_image(buffers, out);
+        self.gpu.copy(buffers.output(), out);
+        let image = self.gpu.map_image(out);
         Ok(image.image())
     }
 
-    /// Renders the image, with a blocking wait to read pixel data from the GPU
-    ///
-    /// This function is only relevant for the web target
-    #[cfg(any(target_arch = "wasm32", doc))]
+    /// Renders the image, with an async wait to read pixel data from the GPU
     pub async fn run_async(
         &self,
         shape: &RenderShape,
         buffers: &mut Buffers,
-        out: &mut ImageReadBuffer,
+        out: &mut ReadBuffer<PixelBufferTag>,
         settings: RenderConfig,
     ) -> Result<Image, SubmitError> {
         self.run_with_vars_async(
@@ -1247,28 +1064,25 @@ impl Context {
         .await
     }
 
-    /// Renders the image, with a blocking wait to read pixel data from the GPU
-    ///
-    /// This function is only relevant for the web target
-    #[cfg(any(target_arch = "wasm32", doc))]
+    /// Renders the image, with an async wait to read pixel data from the GPU
     pub async fn run_with_vars_async(
         &self,
         shape: &RenderShape,
         vars: &ShapeVars<f32>,
         buffers: &mut Buffers,
-        out: &mut ImageReadBuffer,
+        out: &mut ReadBuffer<PixelBufferTag>,
         settings: RenderConfig,
     ) -> Result<Image, SubmitError> {
         self.submit_with_vars(shape, vars, buffers, &settings)?;
-        let image = self.map_image_async(buffers, out).await;
+        self.gpu.copy(buffers.output(), out);
+        let image = self.gpu.map_image_async(out).await;
         Ok(image.image())
     }
 
     /// Submits a single image to be rendered on the GPU
     ///
     /// The resulting image (as a buffer of [`RawDistancePixel`] data) is
-    /// available on the GPU in
-    /// [`buffers.image_storage_buffer()`](Buffers::image_storage_buffer).
+    /// available on the GPU in [`buffers.output()`](Buffers::output).
     pub fn submit(
         &self,
         shape: &RenderShape,
@@ -1371,13 +1185,7 @@ impl Context {
         let mut compute_pass =
             encoder.begin_compute_pass(&wgpu::ComputePassDescriptor {
                 label: None,
-                timestamp_writes: buffers.timestamps.as_ref().map(
-                    |query_set| wgpu::ComputePassTimestampWrites {
-                        query_set,
-                        beginning_of_pass_write_index: Some(0),
-                        end_of_pass_write_index: Some(1),
-                    },
-                ),
+                timestamp_writes: None,
             });
 
         // Build the common config buffer
@@ -1420,97 +1228,6 @@ impl Context {
         // Submit the commands and wait for the GPU to complete
         self.gpu.queue.submit(Some(encoder.finish()));
         Ok(())
-    }
-
-    fn copy_image(&self, buffers: &Buffers, image_out: &mut ImageReadBuffer) {
-        let mut encoder = self.gpu.device.create_command_encoder(
-            &wgpu::CommandEncoderDescriptor {
-                label: Some("map_image"),
-            },
-        );
-        image_out
-            .grow_to_fit(&self.gpu.device, buffers.image_size)
-            .expect(
-                "buffers.image_size should always be \
-                 a valid size for ImageReadBuffer::grow_to_fit",
-            );
-        // Resolve the raw GPU ticks into the resolve buffer, then copy them
-        // into the last 16 bytes of the image buffer
-        if let Some(timestamps) = &buffers.timestamps {
-            encoder.resolve_query_set(timestamps, 0..2, &buffers.ts_buf, 0);
-            encoder.copy_buffer_to_buffer(
-                &buffers.ts_buf,
-                0,
-                image_out.buffer.data(),
-                buffers.pixels.size_bytes(), // offset past the image data
-                buffers.ts_buf.size(),
-            );
-        }
-
-        // Copy from the STORAGE | COPY_SRC -> COPY_DST | MAP_READ buffer
-        encoder.copy_buffer_to_buffer(
-            buffers.pixels.data(),
-            0,
-            image_out.buffer.data(),
-            0,
-            buffers.pixels.size_bytes(),
-        );
-
-        self.gpu.queue.submit(Some(encoder.finish()));
-    }
-
-    /// Synchronously populates and maps a CPU-readable image buffer
-    ///
-    /// The image is borrowed exclusively to avoid double-mapping
-    ///
-    /// This is a blocking function suitable for use on the desktop
-    #[cfg(not(target_arch = "wasm32"))]
-    pub fn map_image<'a>(
-        &self,
-        buffers: &Buffers,
-        image_out: &'a mut ImageReadBuffer,
-    ) -> MappedImage<'a> {
-        self.copy_image(buffers, image_out);
-        let slice = image_out.buffer.map_async(|_| {});
-        self.gpu
-            .device
-            .poll(wgpu::PollType::wait_indefinitely())
-            .unwrap();
-        MappedImage {
-            image: image_out,
-            slice,
-            ns_per_tick: if self.has_timestamps {
-                Some(self.gpu.queue.get_timestamp_period())
-            } else {
-                None
-            },
-        }
-    }
-
-    /// Asynchronously populates and maps a CPU-readable image buffer
-    ///
-    /// The image is borrowed exclusively to avoid double-mapping
-    ///
-    /// This is an `async` function suitable for use in WebAssembly.
-    #[cfg(any(target_arch = "wasm32", doc))]
-    pub async fn map_image_async<'a>(
-        &self,
-        buffers: &Buffers,
-        image_out: &'a mut ImageReadBuffer,
-    ) -> MappedImage<'a> {
-        self.copy_image(buffers, image_out);
-        let (tx, rx) = flume::bounded(0);
-        let slice = image_out.buffer.map_async(move |_| tx.send(()).unwrap());
-        rx.recv_async().await.unwrap();
-        MappedImage {
-            image: image_out,
-            slice,
-            ns_per_tick: if self.has_timestamps {
-                Some(self.gpu.queue.get_timestamp_period())
-            } else {
-                None
-            },
-        }
     }
 }
 
@@ -1663,7 +1380,7 @@ mod test {
             let shape = gpu.shape(&VmShape::from(shape.clone())).unwrap();
             pixel_ctx.submit(&shape, &mut buf, &render_config).unwrap();
             effects_ctx
-                .submit_merge(buf.image_storage_buffer(), true, &mut merge_buf)
+                .submit_merge(buf.output(), true, &mut merge_buf)
                 .unwrap();
         }
         let shape_colors = shapes